+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Gerapy Auto Extractor Changelog
 
+## 0.0.5 (2020-07-20)
+
+### Features
+
+* Add support for `ignoreHTTPSErrors`, `slowMo`, `ignoreDefaultArgs`,
+`handleSIGINT`, `handleSIGTERM`, `handleSIGHUP`, `autoClose` args.
+
 ## 0.0.4 (2020-07-15)
 
 ### Bug Fixes